+++ test/error/unmatched_else_end.sv
@@ -1,0 +1,3 @@
+// pattern: unfinished conditional directive `else started at unmatched_else_end.sv:3:1
+`ifdef FOO
+`else