+++ test/error/unmatched_elsif_end.sv
@@ -1,0 +1,3 @@
+// pattern: unfinished conditional directive `elsif BAR started at unmatched_elsif_end.sv:3:1
+`ifdef FOO
+`elsif BAR