+++ test/error/unmatched_ifdef.sv
@@ -1,1 +1,2 @@
+// pattern: unfinished conditional directive `ifdef FOO started at unmatched_ifdef.sv:2:1
 `ifdef FOO